Print 3: Improved Clips, Top Height, and Square
The clips are now slanted on both sides, so as to hopefully improve the
ability to disassemble the case as well as the length of the clips was
increased to decrease the incident of the angles and decrease chances of
breaking. I also improved the height from 0.22 inches to 0.25 inches in
the lid. Finally, I decreased the square hole to have 3.5mm of clearance
on the top side from the original 5mm clearance.`